+++ gaitmap/utils/array_handling.py
@@ -1,5 +1,5 @@
 """A set of util functions that help to manipulate arrays in any imaginable way."""
-from typing import List, Optional, Tuple
+from typing import List, Optional, Tuple, Iterable, Union, Generator, Iterator
 
 import numba.typed
 import numpy as np
@@ -7,6 +7,8 @@ from numba import njit
 from scipy.interpolate import interp1d
 from scipy.signal import find_peaks
 from typing_extensions import Literal
+
+from gaitmap.utils.datatype_helper import SingleSensorData, SingleSensorStrideList, SingleSensorRegionsOfInterestList
 
 
 def sliding_window_view(arr: np.ndarray, window_length: int, overlap: int, nan_padding: bool = False) -> np.ndarray:
@@ -335,3 +337,37 @@ def _solve_overlap(input_array: np.ndarray, gap_size: int) -> numba.typed.List:
             stack.append(input_array[i])
 
     return stack
+
+
+def iterate_region_data(
+    signal_sequence: Iterable[SingleSensorData],
+    label_sequences: Iterable[Union[SingleSensorStrideList, SingleSensorRegionsOfInterestList]],
+    expected_col_order: Optional[List[str]] = None,
+) -> Iterator[SingleSensorData]:
+    """Iterate over individual strides/ROIs in multiple sensor data sequences.
+
+    This can be helpful if you want to apply operations to each of these strides or ROIs and there is no other way
+    besides iterating over the data.
+
+    Parameters
+    ----------
+    signal_sequence : Iterable[SingleSensorData]
+        A list/iterable of SingleSensorData objects.
+    label_sequences : Iterable[Union[SingleSensorStrideList, SingleSensorRegionsOfInterestList]]
+        A list/iterable of SingleSensorStrideList or SingleSensorRegionsOfInterestList objects.
+        This must have the same length as `signal_sequence` as we expect that the first signal sequence belongs to the
+        first label sequence.
+    expected_col_order : Optional[List[str]]
+        A list of strings that defines the expected column order of the output.
+        This can be helpful to ensure that all output dataframes have the same column order, if it can not be guaranteed
+        that the column order is the same for all input sequences.
+
+    Yields
+    ------
+    SingleSensorData
+        A SingleSensorData object with the data of the next stride/ROI.
+    """
+    for df, labels in zip(signal_sequence, label_sequences):
+        df = df.reindex(columns=expected_col_order)
+        for (_, s, e) in labels[["start", "end"]].itertuples():
+            yield df.iloc[s:e]